Fixed Readering blockAlignment header from Wave

+++ MonoGame.Framework/Audio/SoundEffect.OpenAL.cs
@@ -146,7 +146,7 @@ namespace Microsoft.Xna.Framework.Audio
             var format = BitConverter.ToInt16(header, 0);
             var channels = BitConverter.ToInt16(header, 2);
             var sampleRate = BitConverter.ToInt32(header, 4);
-            var blockAlignment = BitConverter.ToInt32(header, 12);
+            var blockAlignment = BitConverter.ToInt16(header, 12);
 
             // We need to decode MSADPCM.
             var supportsADPCM = OpenALSoundController.GetInstance.SupportsADPCM;
